+++ LottAppl/app/Repositories/Interfaces/LotteryRepositoryInterface.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\Repositories\Interfaces;
+use App\Lottery;
+interface LotteryRepositoryInterface
+{
+    public function getAll();
+
+    public function findById($id);
+
+    public function store($attribute);
+
+    // public function update($id, array $attribute);
+
+    // public function delete($id);
+}
